+++ src/main/java/kr/startoff/backend/domain/user/dto/request/NicknameRequest.java
@@ -1,0 +1,8 @@
+package kr.startoff.backend.domain.user.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class NicknameRequest {
+	String nickname;
+}
