--- src/main/java/kr/startoff/backend/domain/user/dto/request/profile/IntroduceRequest.java
@@ -1,16 +1,0 @@
-package kr.startoff.backend.domain.user.dto.request.profile;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class NicknameAndIntroduceRequest {
-	private String nickname;
-	private String introduce;
-
-	public NicknameAndIntroduceRequest(String nickname, String introduce) {
-		this.nickname = nickname;
-		this.introduce = introduce;
-	}
-}
